Create subarraysSumEqualK
SO, this is a typical question that I've seen multiple times during my assessments and codeathons. 
This is a hashmap + prefixSum question. 
It is solved by the approach that we create a hashmap that stores prefix sum as keys and their count as values. we updated rsum(prefixSum) as we iterate through the loop and check if rsum-k exists as key, if yes, we update the count of key's value or else we map the key in the hashmap with its value, if not present then 1 else if present, then 1+its value.